Fix compilation warnings

+++ panel-plugin/netload.c
@@ -181,7 +181,7 @@ static gboolean update_monitors(gpointer user_data)
         /* correct value to be from 1 ... 100 */
         global->monitor->history[i][0] = net[i];
 
-        if (global->monitor->history[i][0] < 0)
+        if ((gint) global->monitor->history[i][0] < 0)
         {
             global->monitor->history[i][0] = 0;
         }
